Rewrote tournamentHandler to a hook, renamed file and function to useContract
